+++ packages/kraken_animation_player/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
++ Upgrade to kraken 0.9.1
+
+## 1.0.0
+
++ Upgrade to kraken 0.9.0
+
 ## 1.0.0-dev.2
 
 * Migrate to kraken 0.9.0 rc.